backend/requirements.txt
fastapi[standard]
uvicorn[standard]
librosa
tf-explain
lime
shap
tensorflow
numpy
matplotlib
Pillow
scikit-image
opencv-python-headless